Refactored converter classes.
Since ITK to DICOM and DICOM to ITK conversion do not have much in common,
the original class ImageSEGConverter is now splitted into Itk2DicomConverter
and Dicom2ItkConverter.
